refactored documents service, added authorization, seperated documents and files services
