+++ doc-control-app-spring/src/main/java/it/akademija/wizards/controllers/DocumentTypeController.java
@@ -20,10 +20,12 @@ import java.util.List;
 @RestController
 @Api(value = "document types")
 @RequestMapping(value = "/api/doctypes")
+@PreAuthorize("hasRole('ROLE_ADMIN')")
 public class DocumentTypeController {
     @Autowired
     private DocumentTypeService documentTypeService;
 
+    @PreAuthorize("hasRole('ROLE_USER')")
     @ApiOperation(value = "get all document types")
     @RequestMapping(method = RequestMethod.GET)
     @ResponseStatus(HttpStatus.ACCEPTED)
@@ -31,6 +33,7 @@ public class DocumentTypeController {
         return documentTypeService.getDocumentTypes();
     }
 
+    @PreAuthorize("hasRole('ROLE_USER')")
     @ApiOperation(value = "get document type by id")
     @RequestMapping(value="/{id}", method = RequestMethod.GET)
     @ResponseStatus(HttpStatus.ACCEPTED)
